absen-flask/requirements.txt
flask
numpy
pillow
opencv-python
scikit-learn
gunicorn
cmake
dlib
face_recognition
